refactor(codemode): split OpenAPI adapter modules
